+++ example/lib/main.dart
@@ -1,19 +1,20 @@
 import 'package:cubes/cubes.dart';
-import 'package:examplecube/counter/counter_cube.dart';
 import 'package:examplecube/counter/counter_screen.dart';
 import 'package:examplecube/counter/counter_screen_animation.dart';
-import 'package:examplecube/counter_singleton/counter_singleton_cube.dart';
 import 'package:examplecube/counter_singleton/screen_counter_singleton.dart';
 import 'package:examplecube/feedback_manager/feedback_manager_cube.dart';
 import 'package:examplecube/feedback_manager/feedback_manager_screen.dart';
 import 'package:examplecube/pokemon/pokemon_cube.dart';
 import 'package:examplecube/pokemon/pokemon_screen.dart';
 import 'package:examplecube/pokemon/repository/pokemon_repository.dart';
-import 'package:examplecube/text_form_field/text_form_field_cube.dart';
 import 'package:examplecube/text_form_field/text_form_field_screen.dart';
 import 'package:examplecube/todo/todo_cube.dart';
 import 'package:examplecube/todo/todo_list.dart';
 import 'package:flutter/material.dart';
+
+import 'counter/counter_cube.dart';
+import 'counter_singleton/counter_singleton_cube.dart';
+import 'text_form_field/text_form_field_cube.dart';
 
 void main() {
   // register cube
@@ -24,7 +25,6 @@ void main() {
   Cubes.registerDependency((i) => PokemonRepository());
   Cubes.registerDependency((i) => FeedbackManagerCube());
   Cubes.registerDependency((i) => TextFormFieldCube());
-
   runApp(MyApp());
 }
 
@@ -60,52 +60,54 @@ class Home extends StatelessWidget {
         title: Text(Cubes.getString('welcome')),
       ),
       body: Center(
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: <Widget>[
-            RaisedButton(
-              child: Text(Cubes.getString('counter')),
-              onPressed: () {
-                context.goTo(CounterScreen(), settings: RouteSettings(arguments: {'name': 'rafael'}));
-              },
-            ),
-            RaisedButton(
-              child: Text('Counter. Cube in StatefulWidget'),
-              onPressed: () {
-                context.goTo(CounterScreenWithAnimation());
-              },
-            ),
-            RaisedButton(
-              child: Text('Pokemons'),
-              onPressed: () {
-                context.goTo(PokemonScreen());
-              },
-            ),
-            RaisedButton(
-              child: Text(Cubes.getString('singleton')),
-              onPressed: () {
-                context.goTo(ScreenCounterSingleton());
-              },
-            ),
-            RaisedButton(
-              child: Text('Todo'),
-              onPressed: () {
-                context.goTo(TodoList());
-              },
-            ),
-            RaisedButton(
-              child: Text('FeedBackManager'),
-              onPressed: () {
-                context.goTo(FeedbackManagerScreen());
-              },
-            ),
-            RaisedButton(
-              child: Text('CTextFormField'),
-              onPressed: () {
-                context.goTo(TextFormFieldScreen());
-              },
-            ),
-          ],
+        child: SingleChildScrollView(
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: <Widget>[
+              RaisedButton(
+                child: Text(Cubes.getString('counter')),
+                onPressed: () {
+                  context.goTo(CounterScreen(), settings: RouteSettings(arguments: {'name': 'rafael'}));
+                },
+              ),
+              RaisedButton(
+                child: Text('Counter. Cube in StatefulWidget'),
+                onPressed: () {
+                  context.goTo(CounterScreenWithAnimation());
+                },
+              ),
+              RaisedButton(
+                child: Text('Pokemons'),
+                onPressed: () {
+                  context.goTo(PokemonScreen());
+                },
+              ),
+              RaisedButton(
+                child: Text(Cubes.getString('singleton')),
+                onPressed: () {
+                  context.goTo(ScreenCounterSingleton());
+                },
+              ),
+              RaisedButton(
+                child: Text('Todo'),
+                onPressed: () {
+                  context.goTo(TodoList());
+                },
+              ),
+              RaisedButton(
+                child: Text('FeedBackManager'),
+                onPressed: () {
+                  context.goTo(FeedbackManagerScreen());
+                },
+              ),
+              RaisedButton(
+                child: Text('CTextFormField'),
+                onPressed: () {
+                  context.goTo(TextFormFieldScreen());
+                },
+              ),
+            ],
+          ),
         ),
       ),
     );