+++ lib/src/feedback_manager/snackbar_feedback_mixin.dart
@@ -1,57 +1,41 @@
-import 'package:cubes/src/feedback_manager/feedback_manager.dart';
-import 'package:cubes/src/observable/observable_value.dart';
 import 'package:flutter/material.dart';
 
-typedef SnackBar SnackBarByDataBuilder<T>(T data, BuildContext context);
+import '../observable/observable_value.dart';
+import 'feedback_manager.dart';
+
+typedef SnackBarByDataBuilder<T> = SnackBar Function(T data, BuildContext context);
 
 /// Class responsible for configuring the SnackBars
 class CSnackBarController<T> {
-  final SnackBarAction action;
   final ObservableValue<CFeedBackControl<T>> observable;
   final SnackBarByDataBuilder<T> builder;
-  final Duration duration;
-  final ShapeBorder shape;
-  final double elevation;
-  final Color backgroundColor;
-  final EdgeInsetsGeometry margin;
-  final EdgeInsetsGeometry padding;
-  final SnackBarBehavior behavior;
 
+  /// Method used to get SnackBar with data type
   SnackBar doBuild(T data, BuildContext context) {
     return builder(data, context);
   }
 
+  /// Constructor of the CSnackBarController
   CSnackBarController({
     @required this.observable,
     @required this.builder,
-    this.action,
-    this.duration = const Duration(seconds: 4),
-    this.shape,
-    this.elevation,
-    this.backgroundColor,
-    this.margin,
-    this.padding,
-    this.behavior,
   });
 }
 
 /// Mixin responsible for adding listeners to ObservableValue and controlling the display of SnackBars
 mixin SnackBarFeedBackMixin<T extends StatefulWidget> on State<T> {
   List<CSnackBarController> snackBarControllers;
-  Map<CSnackBarController, bool> _mapSnackBarIsShowing = Map();
+  final Map<CSnackBarController, bool> _mapSnackBarIsShowing = {};
 
   /// Configure listeners of the snackBarControllers.
   void confSnackBarFeedBack(List<CSnackBarController> controllers) {
-    this.snackBarControllers = controllers;
-    this.snackBarControllers?.forEach((element) {
-      _mapSnackBarIsShowing[element] = false;
-      element.observable.addListener(() => _listenerDialogController(element));
-    });
+    snackBarControllers = controllers;
+    snackBarControllers?.forEach(_registerSnackBar);
   }
 
   @override
   void dispose() {
-    snackBarControllers?.forEach((element) => element.observable.dispose());
+    snackBarControllers?.forEach(_disposeSnackBar);
     super.dispose();
   }
 
@@ -69,10 +53,22 @@ mixin SnackBarFeedBackMixin<T extends StatefulWidget> on State<T> {
   /// Displays SnackBar with the added settings.
   void _showSnackBar(CSnackBarController element) async {
     _mapSnackBarIsShowing[element] = true;
-    Scaffold.of(context)?.showSnackBar(element.doBuild(element.observable.value.data, context));
-    await Future.delayed(element.duration);
+
+    final snackBar = element.doBuild(element.observable.value.data, context);
+    Scaffold.of(context)?.showSnackBar(snackBar);
+
+    await Future.delayed(snackBar.duration);
+
     _mapSnackBarIsShowing[element] = false;
-    // ignore: invalid_use_of_protected_member
-    element.observable.setInitialValue(element.observable.value.copyWith(show: false));
+    element.observable.setValueWithoutNotify = element.observable.value.copyWith(show: false);
+  }
+
+  void _registerSnackBar(CSnackBarController element) {
+    _mapSnackBarIsShowing[element] = false;
+    element.observable.addListener(() => _listenerDialogController(element));
+  }
+
+  void _disposeSnackBar(CSnackBarController element) {
+    element.observable.dispose();
   }
 }