+++ lib/src/injector/injector.dart
@@ -1,11 +1,12 @@
-import 'package:cubes/src/cube.dart';
+import '../cube.dart';
 
-typedef T CubeInjectorBuilder<T extends Cube>(CInjector injector);
-typedef T DependencyInjectorBuilder<T>(CInjector injector);
+typedef CCubeInjectorBuilder<T extends Cube> = T Function(CInjector injector);
+typedef CDependencyInjectorBuilder<T> = T Function(CInjector injector);
 
+/// Interface responsible to manager dependency injector
 abstract class CInjector {
   /// Method used to register dependency
-  void registerDependency<T>(DependencyInjectorBuilder<T> builder, {String dependencyName, bool isSingleton = false});
+  void registerDependency<T>(CDependencyInjectorBuilder<T> builder, {String dependencyName, bool isSingleton = false});
 
   /// Method used to get dependency
   T getDependency<T>({String dependencyName});