+++ lib/src/localization/cubes_localization.dart
@@ -1,34 +1,40 @@
 import 'dart:async';
 import 'dart:convert';
 
-import 'package:cubes/src/localization/strings_location.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
+import 'strings_location.dart';
+
 /// Responsible for loading the current locate string file
 class CubesLocalization {
+  /// Primary constructor of the  CubesLocalization
   CubesLocalization(this.locale, {this.pathFiles = 'lang/'}) {
-    CStringsLocation.configure(this);
+    CStringsLocation().cubesLocalisation = this;
   }
 
+  /// Locate of the language
   final Locale locale;
+
+  /// path where the json files are located
   final String pathFiles;
 
-  Map<String, String> _sentences;
+  final Map<String, String> _sentences = {};
 
+  /// Load json according to locale
   Future<bool> load() async {
-    String data = await rootBundle.loadString('$pathFiles${this.locale.languageCode}.json');
+    var data = await rootBundle.loadString('$pathFiles${locale.languageCode}.json');
     Map<String, dynamic> _result = json.decode(data);
-
-    this._sentences = new Map();
-    _result.forEach((String key, dynamic value) {
-      this._sentences[key] = value.toString();
+    _sentences.clear();
+    _result.forEach((key, dynamic value) {
+      _sentences[key] = value.toString();
     });
 
     return true;
   }
 
+  /// Get String by key
   String trans(String key) {
-    return this._sentences[key] ?? '';
+    return _sentences[key] ?? '';
   }
 }