+++ lib/src/localization/cubes_localization_delegate.dart
@@ -1,26 +1,32 @@
 import 'dart:async';
 
-import 'package:cubes/src/localization/cubes_localization.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 
-class CubesLocalizationDelegate
-    extends LocalizationsDelegate<CubesLocalization> {
+import 'cubes_localization.dart';
+
+/// Delegate responsible for internationalization from json files
+class CubesLocalizationDelegate extends LocalizationsDelegate<CubesLocalization> {
+  /// primary constructor of the CubesLocalizationDelegate
   const CubesLocalizationDelegate(
     this.supportedLocations, {
     this.pathFiles = 'lang/',
   })  : assert(supportedLocations != null),
         assert(supportedLocations.length > 0);
 
+  /// path where the json files are located
   final String pathFiles;
+
+  /// List of the locations supported
   final List<Locale> supportedLocations;
 
   @override
   bool isSupported(Locale locale) {
-    return true;
+    return supportedLocations.map((e) => e.languageCode).contains(locale.languageCode);
   }
 
+  /// Get delegates to run internationalization
   List<LocalizationsDelegate> get delegates => [
         this,
         DefaultCupertinoLocalizations.delegate,
@@ -30,7 +36,7 @@ class CubesLocalizationDelegate
 
   @override
   Future<CubesLocalization> load(Locale locale) async {
-    CubesLocalization localizations = CubesLocalization(
+    var localizations = CubesLocalization(
       locale,
       pathFiles: pathFiles,
     );