+++ lib/src/observable/observable_value.dart
@@ -2,21 +2,25 @@ import 'package:flutter/foundation.dart';
 
 typedef ModifyValue<T> = T Function(T value);
 
+/// Class that represents our observable
 class ObservableValue<T> extends ChangeNotifier {
   T _value;
   T _lastValue;
 
+  /// Constructor to init ObservableValue with value
   ObservableValue({T value}) {
     _value = value;
   }
 
+  /// Get current value of the observable
   T get value => _value;
+
+  /// Get las value of the observable
   T get lastValue => _lastValue;
 
-  @protected
-  void setInitialValue(T value) {
-    _value = value;
-  }
+  /// Method uses to set value without notify listeners
+  // ignore: avoid_setters_without_getters
+  set setValueWithoutNotify(T value) => _value = value;
 
   void _setValueAndNotify(T newValue) {
     if (T is List) {
@@ -34,7 +38,7 @@ class ObservableValue<T> extends ChangeNotifier {
   void notifyListeners() {
     try {
       super.notifyListeners();
-    } catch (e) {
+    } on Exception catch (_) {
       print('A $runtimeType was used after being disposed');
     }
   }
@@ -43,7 +47,7 @@ class ObservableValue<T> extends ChangeNotifier {
   void dispose() {
     try {
       super.dispose();
-    } catch (e) {
+    } on Exception catch (_) {
       print('Once you have called dispose() on a $runtimeType, it can no longer be used.');
     }
   }