+++ lib/src/util/extensions/observable_value_extensions.dart
@@ -1,8 +1,12 @@
-import 'package:cubes/src/observable/observable_value.dart';
-import 'package:cubes/src/widgets/observer.dart';
+import 'package:cubes/cubes.dart';
 import 'package:flutter/widgets.dart';
 
+import '../../observable/observable_value.dart';
+import '../../widgets/observer.dart';
+
+/// Extension to facilitate the use of CObserver
 extension ObservableValueExtensions on ObservableValue {
+  /// Use to create CObserver by ObservableValue
   CObserver build<T>(
     ObserverBuilder<T> build, {
     bool animate = false,
@@ -19,4 +23,16 @@ extension ObservableValueExtensions on ObservableValue {
       builder: build,
     );
   }
+}
+
+/// Extension to facilitate create ObservableValue
+extension ObservableValueExtension<T> on T {
+  /// Create ObservableValue<T>
+  ObservableValue<T> get obsValue => ObservableValue<T>(value: this);
+}
+
+/// Extension to facilitate create ObservableList
+extension ListToObservableValue<T> on List<T> {
+  /// Create ObservableList<T>
+  ObservableList<T> get obsValue => ObservableList<T>(value: this);
 }