+++ test/widgets/observer/observer_test.dart
@@ -3,20 +3,21 @@ import 'package:flutter_test/flutter_test.dart';
 import 'observer_robot.dart';
 
 void main() {
-  testWidgets('Should render with initial value', (WidgetTester tester) async {
+  testWidgets('Should render with initial value', (tester) async {
     final robot = ObserverRobot(tester);
     await robot.setup();
     await robot.assetInitialValue();
   });
 
-  testWidgets('Should update render with new value', (WidgetTester tester) async {
+  testWidgets('Should update render with new value', (tester) async {
     final robot = ObserverRobot(tester);
     await robot.setup();
     robot.changeValueWithUpdate(25);
     await robot.assetValue(25);
   });
 
-  testWidgets('Should update render with new value using modify', (WidgetTester tester) async {
+  // ignore: lines_longer_than_80_chars
+  testWidgets('Should update render with new value using modify', (tester) async {
     final robot = ObserverRobot(tester);
     await robot.setup();
     robot.incrementValueWithModify(25);
@@ -24,10 +25,10 @@ void main() {
     await robot.assetValue(30);
   });
 
-  testWidgets('Should note update render when not accept in WhenBuild', (WidgetTester tester) async {
+  testWidgets('Should note update render when not accept in WhenBuild', (tester) async {
     final robot = ObserverRobot(tester);
     await robot.setup();
-    robot.registerWhenBuild((lastV, newV) {
+    robot.whenBuild = ((lastV, newV) {
       return (newV - lastV) == 1;
     });
     robot.changeValueWithUpdate(5);